+++ js/jquery.terminal.js
@@ -4,7 +4,7 @@
  *  __ / // // // // // _  // _// // / / // _  // _//     // //  \/ // _ \/ /
  * /  / // // // // // ___// / / // / / // ___// / / / / // // /\  // // / /__
  * \___//____ \\___//____//_/ _\_  / /_//____//_/ /_/ /_//_//_/ /_/ \__\_\___/
- *           \/              /____/                              version 2.2.0
+ *           \/              /____/                              version DEV
  *
  * This file is part of jQuery Terminal. https://terminal.jcubic.pl
  *
@@ -35,7 +35,7 @@
  * emoji regex v7.0.1 by Mathias Bynens
  * MIT license
  *
- * Date: Sun, 17 Feb 2019 12:24:21 +0000
+ * Date: Sat, 16 Mar 2019 14:41:43 +0000
  */
 /* global location, setTimeout, window, global, sprintf, setImmediate,
           IntersectionObserver,  ResizeObserver, module, require, define,
@@ -1376,10 +1376,12 @@
                 no_keydown = true;
 
                 self.set('');
+                var promise;
                 if (settings.commands) {
-                    settings.commands.call(self, tmp);
+                    promise = settings.commands.call(self, tmp);
                 }
-                if (is_function(prompt)) {
+                if (is_function(prompt) &&
+                    (promise && promise.isResolved() || !promise)) {
                     draw_prompt();
                 }
                 clip.val('');
@@ -3460,8 +3462,8 @@
     }
     // -------------------------------------------------------------------------
     $.terminal = {
-        version: '2.2.0',
-        date: 'Sun, 17 Feb 2019 12:24:21 +0000',
+        version: 'DEV',
+        date: 'Sat, 16 Mar 2019 14:41:43 +0000',
         // colors from https://www.w3.org/wiki/CSS/Properties/color/keywords
         color_names: [
             'transparent', 'currentcolor', 'black', 'silver', 'gray', 'white',