+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 2.2.1
 ### Bugfix
 * fix selecting whole text of single .cmd line
+* fix double call of async prompt with async interpreter [#474](https://github.com/jcubic/jquery.terminal/issues/474)
 
 ## 2.2.0
 ### Features